Remove useless utils
