--- common/src/main/kotlin/org/jetbrains/research/common/gumtree/ReflectionUtils.kt
@@ -1,56 +1,0 @@
-package org.jetbrains.research.common.gumtree
-
-import com.github.gumtreediff.actions.model.Action
-import com.github.gumtreediff.actions.model.Insert
-import com.github.gumtreediff.actions.model.Move
-import com.github.gumtreediff.tree.ITree
-import com.intellij.psi.PsiElement
-import com.jetbrains.python.psi.PyElement
-import java.lang.reflect.Method
-import java.lang.reflect.Modifier
-
-sealed class PyElementChild
-class EmptyChildContainer : PyElementChild()
-class OneChildContainer(val element: PyElement) : PyElementChild()
-class ManyChildrenContainer(val elements: Array<PyElement>) : PyElementChild()
-
-internal fun extractChildrenByFieldName(element: PyElement): Map<String, PyElementChild> {
-    val psiChildren: Set<PsiElement> = element.children.toSet()
-    val methods: List<Method> = element.javaClass.methods
-        .filter { Modifier.isPublic(it.modifiers) }
-        .filter {
-            PyElement::class.java.isAssignableFrom(it.returnType)
-                    || Array<PyElement>::class.java.isAssignableFrom(it.returnType)
-        }
-    val childByMethodName = hashMapOf<String, PyElementChild>()
-    for (method in methods) {
-        val potentialChild = try {
-            method.invoke(element)
-        } catch (ex: Exception) {
-            continue
-        }
-        if (potentialChild is PyElement && psiChildren.contains(potentialChild)) {
-            childByMethodName[method.name] = OneChildContainer(potentialChild)
-        } else if (potentialChild is Array<*>
-                && potentialChild.all { it is PyElement && psiChildren.contains(it) }
-        ) {
-            childByMethodName[method.name] = ManyChildrenContainer(potentialChild as Array<PyElement>)
-        } else {
-            childByMethodName[method.name] = EmptyChildContainer()
-        }
-    }
-    return childByMethodName
-}
-
-fun getAllTreesFromActions(actions: Collection<Action>): List<ITree> {
-    val result = arrayListOf<ITree>()
-    for (action in actions) {
-        result.add(action.node)
-        if (action is Insert) {
-            result.add(action.parent)
-        } else if (action is Move) {
-            result.add(action.parent)
-        }
-    }
-    return result
-}
